# backend/insomnia/strobe-openapi.yaml
openapi: 3.0.3
info:
  title: Strobe Local API
  version: 1.0.0
  contact:
    name: CAB432 Teaching Team
    email: cab432@qut.edu.au
  description: |
    Local OpenAPI specification for the Strobe server.

    This spec is designed for direct import into Insomnia or other API clients and reflects
    the request and response shapes implemented by the Express server in `server/src`.

    Authentication uses JWT bearer tokens. Register or log in first, then send:

    `Authorization: Bearer <token>`
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT token only. Do not include the "Bearer " prefix.
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000
...
    QueryLimit:
      in: query
      name: limit
      required: false
      description: Maximum number of records to return. The server caps this at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Title is required
...
    UserStats:
      type: object
      required: [followers, following, posts]
      properties:
        followers:
          type: integer
          example: 12
        following:
          type: integer
          example: 7
        posts:
          type: integer
          example: 4
    User:
      type: object
      required: [id, username, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake_bradford
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        role:
          type: string
          enum: [user, moderator]
          example: user
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
    EnrichedUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [stats, currentUserFollows]
          properties:
            stats:
              $ref: "#/components/schemas/UserStats"
            currentUserFollows:
              type: boolean
              example: false
...
    UserResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
          example: Success
        user:
          $ref: "#/components/schemas/EnrichedUser"
    UserListResponse:
      type: object
      required: [message, users]
      properties:
        message:
          type: string
          example: Success
        users:
          type: array
          items:
            $ref: "#/components/schemas/EnrichedUser"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Deleted successfully
  responses:
...
    Unauthorised:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorised
            message: Invalid or expired token
    Forbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Forbidden
            message: You can only modify your own posts
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NotFound
            message: Post not found
paths:
...
  /v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: |
        Returns public user profiles. If a bearer token is supplied, each user also includes
        currentUserFollows based on the authenticated caller.
      parameters:
        - in: query
          name: q
          required: false
          description: Optional case-insensitive username search fragment.
          schema:
            type: string
            example: alex
        - $ref: "#/components/parameters/QueryLimit"
      responses:
        "200":
          description: Matching users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
              example:
                message: Success
                users:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake_bradford
                    email: jake.bradford@example.com
                    role: user
                    createdAt: 2026-07-08T09:00:00.000Z
                    updatedAt: 2026-07-08T09:00:00.000Z
                    stats:
                      followers: 12
                      following: 7
                      posts: 4
                    currentUserFollows: false
  /v1/users/{id}:
    get:
      tags: [Users]
      operationId: getUserProfile
      summary: Get user profile
      description: Returns one public user profile with follower, following, and post counts.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                message: Success
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake_bradford
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                  stats:
                    followers: 12
                    following: 7
                    posts: 4
                  currentUserFollows: false
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Users]
      operationId: deleteOwnAccount
      summary: Delete own account
      description: Deletes the authenticated user's own account and cascades related posts, comments, likes, follows, moments, and uploads.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...